Make Directory#copyFrom abstract (#14985)
